linux-gen: modular-framework: move buffer subsystem header to a public folder
Move buffer subsystem header to a public folder.
