+++ include/Makefile.am
@@ -68,6 +68,10 @@ odpdrvspecinclude_HEADERS = \
 		  odp/drv/spec/std_types.h \
 		  odp/drv/spec/sync.h
 
+subsystemspecincludedir= $(includedir)/subsystem/spec
+subsystemspecinclude_HEADERS = \
+		  subsystem/spec/buffer_subsystem.h
+
 odpapiabidefaultincludedir= $(includedir)/odp/arch/default/api/abi
 odpapiabidefaultinclude_HEADERS = \
 	odp/arch/default/api/abi/buffer.h \
