+++ platform/linux-dpdk/buffer/dpdk.c
@@ -9,7 +9,7 @@
 #include <odp/api/buffer.h>
 #include <odp_buffer_internal.h>
 #include <odp_buffer_inlines.h>
-#include <odp_buffer_subsystem.h>
+#include <subsystem/spec/buffer_subsystem.h>
 #include <odp_debug_internal.h>
 #include <odp_pool_internal.h>
