+++ platform/linux-generic/Makefile.am
@@ -160,7 +160,6 @@ noinst_HEADERS = \
 		  include/odp_bitmap_internal.h \
 		  include/odp_bitset.h \
 		  include/odp_buffer_internal.h \
-		  include/odp_buffer_subsystem.h \
 		  include/odp_classification_datamodel.h \
 		  include/odp_classification_inlines.h \
 		  include/odp_classification_internal.h \
